refactoring test to move testdata to a seperate file, so source code tokenizers stop freaking out.
