+++ test/Initializer.js
@@ -7,6 +7,7 @@ const TxPermission = artifacts.require('TxPermissionHbbft');
 const Certifier = artifacts.require('CertifierHbbft');
 const KeyGenHistory = artifacts.require('KeyGenHistory');
 const Initializer =  artifacts.require('InitializerHbbft');
+const testdata = require('./testhelpers/data');
 
 const BN = web3.utils.BN;
 
@@ -56,20 +57,7 @@ contract('InitializerHbbft', async accounts => {
      '0x20200000000000000000000000000000', 
      '0x30300000000000000000000000000000'];
 
-  
-  const partsRaw =  [[0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,162,238,181,164,119,226,238,142,43,227,179,139,95,25,108,237,11,190,84,64,234,22,245,210,187,24,142,145,31,26,197,119,92,171,182,30,182,191,151,218,208,96,103,110,14,170,189,220,3,0,0,0,0,0,0,0,153,0,0,0,0,0,0,0,4,21,73,242,168,17,236,103,137,92,198,188,151,185,9,174,175,170,174,57,210,222,4,81,210,164,8,69,148,77,228,38,240,153,232,223,247,87,154,182,186,84,25,66,211,40,191,201,51,82,23,210,22,213,222,200,8,206,218,143,128,49,28,60,223,37,52,95,47,33,245,72,158,241,216,97,244,234,136,89,239,199,11,252,232,95,18,174,135,13,237,240,213,206,206,85,52,124,247,123,199,87,112,132,42,136,94,223,15,75,137,122,27,20,92,167,40,222,99,160,147,245,251,242,221,134,74,193,248,249,246,54,58,185,109,78,212,36,101,241,65,82,183,246,62,156,217,247,255,92,205,160,220,153,0,0,0,0,0,0,0,4,249,64,230,166,97,82,165,124,79,27,92,167,19,229,52,40,203,120,236,87,142,125,122,227,174,162,64,228,170,35,46,169,74,119,13,26,126,67,141,151,3,86,253,135,156,100,60,42,68,245,201,4,66,216,34,29,88,229,82,88,171,213,215,25,96,245,193,33,229,228,139,233,239,14,164,187,39,151,82,164,72,181,215,161,64,167,181,169,96,89,32,93,180,176,201,116,8,190,99,179,132,43,11,107,59,223,207,173,43,197,181,169,127,184,189,122,220,159,133,133,17,192,5,149,159,97,197,143,104,103,252,93,120,213,56,242,193,4,27,187,151,60,172,129,55,69,23,243,110,119,16,151,153,0,0,0,0,0,0,0,4,144,226,236,244,137,245,236,219,56,172,227,38,22,176,15,182,184,184,45,175,169,208,170,150,182,129,76,191,47,255,123,213,136,70,214,129,234,24,96,55,181,159,73,123,96,55,238,104,140,106,39,86,194,82,217,229,245,194,114,78,33,192,217,137,147,141,230,183,211,227,236,196,102,226,30,247,198,211,117,159,70,135,219,249,232,158,39,102,214,42,163,97,67,172,10,15,53,212,46,69,144,227,87,81,113,248,238,229,74,78,4,158,139,125,42,181,220,208,108,188,71,4,193,66,135,63,246,180,166,251,117,160,133,3,14,154,17,254,34,215,98,212,107,3,219,145,162,81,154,129,98,30],[0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,178,73,138,14,3,78,162,102,191,223,232,10,81,96,118,193,17,245,152,100,177,160,138,30,241,144,254,80,221,191,133,251,43,252,251,55,111,234,36,30,11,227,3,209,11,29,97,12,3,0,0,0,0,0,0,0,153,0,0,0,0,0,0,0,4,190,190,139,180,190,148,3,35,23,114,210,184,188,37,29,84,197,247,58,199,180,139,151,50,108,72,160,72,104,120,177,113,88,114,102,35,83,214,49,107,226,189,62,146,134,227,121,204,221,170,229,103,242,35,66,174,12,222,96,244,80,109,177,162,22,173,241,253,188,107,213,119,32,142,210,120,23,6,255,53,195,183,16,224,211,157,34,76,41,196,121,60,10,124,231,193,177,178,201,176,242,10,34,30,74,16,76,255,101,225,104,179,5,15,231,147,247,214,103,172,127,161,226,217,18,140,93,42,235,252,105,19,214,191,80,37,230,77,82,122,248,38,124,30,68,164,81,129,51,162,87,154,153,0,0,0,0,0,0,0,4,45,167,32,19,176,115,181,244,240,144,71,62,123,234,47,139,101,220,132,48,31,222,255,194,15,65,214,103,75,68,161,78,40,143,28,242,143,191,87,76,169,19,24,217,145,231,90,139,69,34,90,12,175,146,173,49,116,34,148,58,62,142,168,154,1,79,76,99,59,135,153,22,213,9,80,206,179,110,110,123,8,42,101,53,36,241,202,64,203,169,148,175,0,2,217,163,109,89,248,214,243,66,44,69,11,100,90,132,74,218,16,240,20,157,154,134,8,63,131,18,59,157,105,93,133,47,166,97,196,160,48,54,218,118,46,146,224,36,123,226,204,171,215,17,73,92,147,55,1,41,54,32,153,0,0,0,0,0,0,0,4,116,101,44,65,60,102,230,141,166,221,113,19,43,220,253,67,27,109,103,171,33,42,189,108,84,229,139,200,63,218,76,180,101,159,208,102,117,168,34,185,211,61,12,143,250,49,210,194,39,135,64,150,174,159,196,254,136,138,229,53,214,85,228,119,125,222,201,207,137,10,57,88,233,155,50,94,50,238,174,188,71,119,14,67,50,222,24,92,82,87,126,120,5,30,112,40,106,216,120,181,16,197,44,52,86,129,181,241,249,91,31,253,24,42,106,14,5,175,226,29,226,200,218,97,99,81,125,219,133,214,228,198,138,241,41,193,138,218,51,168,61,46,35,86,144,153,99,180,150,11,4,112],[0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,170,254,25,182,81,8,235,232,45,133,72,208,54,10,151,224,33,250,175,154,155,189,139,187,155,22,199,114,11,221,65,195,66,45,57,81,129,41,252,72,173,189,241,122,193,176,97,209,3,0,0,0,0,0,0,0,153,0,0,0,0,0,0,0,4,106,140,219,6,75,230,222,99,131,82,163,144,68,231,71,69,7,214,27,149,103,189,152,156,32,231,59,113,140,92,46,178,144,219,172,255,183,225,111,161,65,244,242,135,6,223,69,149,189,167,16,27,107,145,136,2,12,12,243,105,75,232,172,184,119,164,14,223,101,219,182,114,55,189,158,72,194,84,185,46,74,77,128,16,2,195,132,20,78,189,141,196,64,145,81,196,15,182,249,21,181,205,247,242,35,208,150,216,84,254,243,5,232,211,110,196,137,93,18,176,42,201,176,20,114,94,139,48,60,113,173,246,59,152,54,204,22,124,81,29,86,57,176,80,68,85,253,54,169,193,165,91,153,0,0,0,0,0,0,0,4,104,82,218,8,57,188,206,90,232,210,234,64,176,216,102,167,61,82,75,197,47,186,148,82,186,148,182,186,30,33,232,250,130,0,204,228,25,63,7,21,157,244,225,15,53,251,119,149,11,45,150,165,213,162,59,248,44,174,177,88,128,38,11,228,110,198,243,237,199,79,167,252,11,250,131,140,109,122,89,181,115,128,198,25,153,196,196,248,72,212,208,54,32,102,7,208,238,103,32,125,62,111,222,83,6,5,148,59,10,74,29,64,4,251,189,34,108,226,206,8,77,109,181,121,208,54,82,193,89,132,22,238,147,160,28,88,39,32,50,147,222,138,153,191,243,96,34,218,212,51,159,23,153,0,0,0,0,0,0,0,4,20,112,232,10,136,250,15,177,137,89,249,144,240,199,202,172,135,49,130,254,98,208,175,191,151,185,180,168,51,4,128,163,66,75,212,151,21,75,144,199,45,191,89,166,90,109,221,74,10,97,171,7,180,219,192,127,71,90,221,87,223,233,110,232,90,245,107,197,138,136,222,206,16,139,147,205,107,117,53,190,36,204,74,112,199,210,123,131,71,199,95,6,181,165,60,186,116,116,46,203,91,155,5,187,22,95,215,92,2,77,79,204,34,75,118,213,76,78,37,180,26,63,220,229,190,164,60,155,245,136,80,165,197,44,64,108,99,152,129,80,83,235,94,234,233,252,49,252,182,182,178,2]]
-
-  const acks =  [[[0,0,0,0,0,0,0,0,3,0,0,0,0,0,0,0,145,0,0,0,0,0,0,0,4,49,56,206,120,208,7,197,245,119,88,166,31,191,100,103,74,94,140,184,31,111,69,189,192,100,241,203,205,189,66,172,233,179,112,132,219,243,78,177,183,108,34,50,65,27,107,63,176,91,62,117,81,36,102,5,36,26,45,216,19,213,65,150,109,205,225,150,94,35,42,219,196,29,193,159,197,151,197,236,146,233,109,92,114,132,158,253,177,201,2,138,19,105,71,220,139,123,226,74,75,79,151,109,146,250,223,71,240,101,163,100,232,34,18,27,190,42,160,240,234,71,10,190,225,42,86,126,84,180,104,163,30,81,152,217,249,46,39,75,180,37,174,25,146,145,0,0,0,0,0,0,0,4,209,118,244,76,61,65,111,73,69,38,163,50,120,117,116,225,140,206,177,188,80,71,117,111,112,177,119,120,97,192,129,63,57,10,114,60,228,137,64,189,13,184,18,147,153,76,28,166,129,11,39,164,61,116,170,155,126,161,25,67,121,213,137,1,248,133,179,141,36,108,52,127,4,2,206,60,231,133,30,94,38,184,209,145,138,230,43,82,129,228,215,42,47,192,240,126,84,97,60,129,195,203,232,81,192,119,161,187,1,255,141,201,188,35,19,205,98,234,206,113,235,15,84,88,223,168,79,142,139,113,138,109,198,168,166,134,84,148,62,238,107,210,58,184,145,0,0,0,0,0,0,0,4,169,49,88,116,160,40,158,158,37,164,106,95,107,236,212,41,98,167,202,51,165,245,162,176,162,13,27,94,205,72,164,85,128,234,178,50,240,91,53,79,54,43,208,184,144,253,54,115,5,143,67,187,222,154,181,174,39,92,120,202,117,148,241,147,163,105,150,75,56,172,215,31,143,93,13,100,27,25,52,111,206,176,55,61,33,42,77,189,94,69,182,212,97,111,43,129,38,216,218,117,64,250,128,254,104,165,160,72,205,2,142,157,116,95,218,49,143,85,160,190,195,205,97,21,77,85,58,4,94,3,253,167,237,55,75,101,200,198,97,122,46,153,85,104],[1,0,0,0,0,0,0,0,3,0,0,0,0,0,0,0,145,0,0,0,0,0,0,0,4,23,51,20,34,33,38,160,91,243,39,211,84,188,187,32,88,121,194,65,132,6,33,255,93,159,16,49,126,232,121,180,244,229,106,87,125,60,156,166,70,125,207,203,103,115,212,153,133,130,108,248,22,122,230,95,106,74,90,123,120,171,33,102,115,12,56,14,171,238,173,226,146,136,62,60,201,28,178,216,162,107,163,48,76,227,28,104,85,33,252,28,249,108,249,27,72,213,44,161,33,253,20,154,127,29,27,84,76,10,232,211,60,248,37,50,100,210,125,71,109,155,165,246,47,85,219,71,51,28,163,219,240,253,113,156,131,70,148,248,187,146,135,167,165,145,0,0,0,0,0,0,0,4,192,60,165,5,196,71,96,49,93,153,135,242,152,67,254,94,244,186,64,132,128,112,67,102,177,240,181,105,170,39,57,176,247,119,207,134,253,249,122,30,151,118,91,107,16,88,156,112,214,181,199,115,246,190,173,85,153,49,13,107,64,173,3,76,236,221,54,120,163,47,192,242,19,35,90,153,37,202,166,22,33,252,250,218,68,80,197,134,147,230,72,107,239,119,69,53,177,53,92,248,15,154,177,116,202,54,140,118,62,188,103,36,5,160,74,145,223,101,251,190,96,122,191,34,227,47,66,152,165,238,80,210,237,104,71,158,119,13,105,141,144,0,178,141,145,0,0,0,0,0,0,0,4,225,41,255,228,45,231,42,88,201,191,196,69,54,108,214,0,203,113,53,203,41,155,102,29,59,80,28,21,199,78,156,141,205,214,137,150,105,116,44,164,70,53,174,153,241,100,139,87,176,6,39,38,21,0,186,0,126,67,56,61,179,90,176,124,223,224,160,121,228,119,98,103,241,178,118,96,137,223,193,235,135,196,83,1,81,221,201,242,238,155,66,12,221,16,213,18,198,232,69,253,82,230,12,212,83,137,100,18,169,241,101,21,135,126,201,168,200,16,209,248,169,108,55,250,131,104,64,209,203,142,56,15,250,225,253,236,25,208,180,120,9,110,234,184],[2,0,0,0,0,0,0,0,3,0,0,0,0,0,0,0,145,0,0,0,0,0,0,0,4,100,13,69,238,212,63,79,226,98,251,16,122,85,165,126,121,80,111,42,109,158,70,24,38,38,111,158,79,150,173,105,34,161,225,200,179,88,218,70,23,237,26,224,110,79,152,70,59,118,103,5,164,71,141,68,50,222,235,190,144,52,217,236,70,228,228,120,58,4,140,229,232,184,37,120,151,84,2,188,45,166,166,161,187,131,248,99,81,88,35,244,114,124,35,188,35,226,110,116,96,22,200,7,218,13,242,120,227,18,75,233,181,235,133,22,176,183,146,218,62,35,32,39,72,51,200,142,43,139,0,30,41,253,158,80,250,68,109,221,189,177,28,230,134,145,0,0,0,0,0,0,0,4,230,16,43,212,214,93,133,200,66,103,14,32,84,113,112,209,67,34,129,26,136,225,228,116,249,229,195,82,36,247,52,191,165,171,143,36,78,7,95,249,100,75,104,11,108,10,125,197,38,196,120,79,205,57,54,20,100,210,169,117,30,216,101,211,1,65,202,100,146,162,231,241,5,28,237,106,82,140,136,157,60,156,166,30,139,215,218,54,192,91,26,101,227,29,219,61,209,239,219,79,194,67,158,79,171,209,46,25,90,148,156,228,253,187,118,109,190,207,97,190,225,163,65,76,250,34,125,135,80,183,141,21,182,39,53,50,53,17,231,237,74,203,174,220,145,0,0,0,0,0,0,0,4,49,229,95,158,235,69,183,126,193,92,120,197,77,79,151,132,150,130,123,226,27,226,253,204,82,95,181,114,206,239,142,232,86,117,230,232,39,123,255,21,227,52,86,154,154,217,229,4,14,111,147,103,147,193,202,157,166,220,177,1,38,93,150,126,205,238,236,213,7,187,40,13,219,11,219,121,36,90,202,159,30,159,62,6,87,109,210,169,62,53,162,10,41,238,83,97,148,23,65,146,115,98,52,29,241,192,189,6,150,174,21,123,176,123,226,75,128,104,115,68,6,217,1,159,67,121,20,198,240,72,151,231,217,154,29,120,64,123,232,14,202,99,231,251]],[[0,0,0,0,0,0,0,0,3,0,0,0,0,0,0,0,145,0,0,0,0,0,0,0,4,184,101,147,79,192,143,119,143,72,141,194,218,23,81,123,6,94,214,88,150,139,211,68,135,97,69,29,94,41,91,26,204,61,24,91,217,137,202,97,9,18,148,103,186,213,187,223,230,25,110,175,110,90,114,194,73,165,219,155,212,227,186,187,166,92,183,23,67,2,63,180,17,40,8,93,62,100,159,223,39,225,254,172,23,104,91,22,115,192,67,106,71,6,46,27,2,222,129,174,89,183,59,187,198,112,18,17,54,82,144,204,17,210,245,97,214,176,33,220,17,74,18,178,66,43,250,173,88,181,106,91,105,221,54,26,223,99,64,50,20,108,139,245,104,145,0,0,0,0,0,0,0,4,103,215,2,70,67,85,254,117,92,62,181,133,48,134,52,71,18,106,60,87,218,85,75,223,233,176,92,120,124,200,214,185,85,234,88,94,29,171,232,194,5,189,88,126,173,116,216,3,98,168,191,124,33,154,219,4,137,156,101,110,139,4,173,100,212,241,251,197,120,8,220,188,12,228,174,113,167,127,4,97,115,155,57,38,93,245,5,3,114,61,38,194,193,33,48,196,239,88,116,63,24,255,62,246,73,77,234,138,136,93,133,105,9,29,172,52,46,101,13,129,147,75,39,25,160,179,25,88,185,85,81,240,130,180,145,12,165,206,127,142,31,233,182,117,145,0,0,0,0,0,0,0,4,125,157,244,235,22,158,28,67,94,19,16,164,238,104,224,161,47,190,216,231,249,164,129,25,238,142,145,219,48,178,181,190,209,90,107,218,157,54,101,107,81,249,140,245,169,31,47,89,39,165,99,119,87,179,112,49,81,69,2,25,227,202,37,246,115,211,85,250,25,127,52,13,228,153,108,11,47,65,133,2,45,120,125,202,166,126,113,98,185,21,91,152,3,189,240,11,253,38,225,249,239,93,239,187,102,106,145,78,99,42,15,190,254,157,170,114,170,142,108,108,67,172,246,3,209,238,143,155,254,156,246,121,30,86,133,136,214,44,31,156,211,224,222,237],[1,0,0,0,0,0,0,0,3,0,0,0,0,0,0,0,145,0,0,0,0,0,0,0,4,245,37,118,90,114,11,69,85,13,113,7,10,46,220,128,237,70,201,68,71,248,182,116,96,130,201,35,161,178,141,45,55,39,87,44,21,118,135,5,60,194,1,216,148,24,127,129,234,172,95,202,131,157,77,245,198,63,183,33,45,123,242,220,247,20,196,159,155,204,250,69,161,10,93,189,70,10,97,122,141,22,61,77,21,49,86,64,137,21,43,221,154,68,7,38,203,241,208,207,125,144,211,255,135,247,27,211,203,6,189,88,198,161,156,15,58,244,254,58,171,120,239,1,83,22,236,178,245,156,59,10,11,21,33,121,227,47,25,62,160,193,73,252,123,145,0,0,0,0,0,0,0,4,50,105,164,252,224,42,30,14,83,98,248,84,249,220,134,92,208,38,53,131,214,224,87,147,147,185,212,9,134,64,43,50,28,108,234,96,14,107,243,170,216,151,162,163,208,7,198,253,12,236,16,203,174,4,48,135,216,146,175,253,10,35,203,153,123,20,105,163,223,250,166,82,141,225,134,94,27,124,157,123,110,174,31,240,203,25,146,16,101,147,74,70,107,68,78,61,178,42,126,11,177,157,18,109,54,194,124,160,236,96,187,113,162,0,68,209,93,80,239,239,73,177,74,163,97,248,138,157,21,225,77,71,152,208,192,109,19,49,222,147,105,230,187,26,145,0,0,0,0,0,0,0,4,196,48,14,64,241,20,121,45,118,108,242,220,188,251,9,107,149,237,200,167,223,80,201,104,246,181,10,255,48,238,65,55,250,9,224,197,155,166,88,92,243,71,134,100,103,55,215,124,201,181,48,208,69,30,86,179,107,38,74,80,138,169,30,221,251,16,192,12,161,130,59,65,116,91,186,37,161,54,141,67,221,161,216,44,175,218,171,197,177,244,193,164,192,140,196,90,171,162,210,100,194,164,195,189,249,124,97,79,250,133,185,89,174,62,150,155,35,87,33,82,61,38,232,43,228,126,92,91,225,253,240,28,91,77,191,108,152,75,154,244,98,221,191,41],[2,0,0,0,0,0,0,0,3,0,0,0,0,0,0,0,145,0,0,0,0,0,0,0,4,208,125,51,24,222,224,29,26,2,119,35,43,81,99,219,140,237,96,92,48,25,35,74,4,122,119,52,55,237,99,4,144,36,224,73,174,186,0,12,39,68,103,42,134,94,0,103,107,106,211,95,98,195,248,225,183,114,52,180,10,14,36,107,238,92,93,0,121,27,201,235,125,8,100,243,73,198,214,4,217,47,60,70,194,237,149,25,135,181,31,69,97,89,42,189,55,78,56,5,104,213,119,91,115,37,201,144,98,167,187,33,253,113,207,76,20,178,24,214,144,89,186,166,135,14,51,107,97,144,163,139,211,32,17,212,27,207,155,228,3,232,99,29,60,145,0,0,0,0,0,0,0,4,234,91,111,181,233,101,114,213,130,88,131,151,254,87,194,245,250,230,67,44,21,75,113,99,35,227,123,228,175,10,249,92,34,132,237,153,77,82,121,49,55,152,228,11,46,181,146,34,5,3,138,34,124,85,0,224,9,106,36,210,163,56,94,235,72,252,235,77,30,179,248,61,123,143,39,120,163,112,18,35,141,62,143,237,22,9,66,21,3,62,100,5,46,255,142,248,160,235,218,215,209,249,65,40,221,198,201,113,8,223,82,78,41,189,237,58,52,36,237,29,112,210,54,79,232,40,100,68,151,113,155,194,127,109,170,203,56,80,214,156,160,169,170,185,145,0,0,0,0,0,0,0,4,144,88,183,137,94,77,106,60,58,106,25,131,143,146,102,70,96,101,51,196,134,213,70,185,103,71,239,116,121,157,246,50,143,242,178,108,249,215,239,222,147,243,250,216,14,192,52,246,247,244,232,119,186,159,2,36,114,163,241,82,32,197,173,98,98,42,126,14,241,129,221,106,67,41,115,178,175,34,150,92,28,40,189,70,225,15,81,134,167,214,149,173,118,25,166,226,7,163,125,124,193,175,36,163,136,128,131,86,62,64,56,176,251,4,91,42,46,167,97,141,136,49,137,96,99,124,173,113,76,33,36,202,116,5,0,254,78,198,243,113,77,124,227,174]],[[0,0,0,0,0,0,0,0,3,0,0,0,0,0,0,0,145,0,0,0,0,0,0,0,4,23,250,79,243,54,117,62,80,179,97,221,176,136,63,222,209,236,75,87,159,246,168,252,213,57,22,96,252,27,160,103,102,80,93,78,34,182,54,109,189,204,137,210,232,139,103,3,147,219,189,217,31,68,104,247,120,169,69,79,85,180,196,118,81,189,161,132,143,123,51,43,132,220,79,82,247,160,134,243,153,40,54,207,158,51,52,168,105,127,210,129,206,119,10,251,112,225,3,181,153,204,171,86,17,102,145,81,48,60,27,219,84,8,200,15,236,158,110,58,86,225,242,33,43,29,211,190,141,197,77,250,211,84,108,189,249,186,15,248,239,126,130,47,1,145,0,0,0,0,0,0,0,4,99,232,72,81,216,83,82,177,113,154,101,171,191,102,16,53,2,147,223,165,79,243,190,113,10,157,206,135,222,5,217,65,54,23,177,200,142,189,10,163,50,96,137,158,37,217,106,216,161,158,201,42,218,173,255,223,107,187,10,160,246,209,49,119,13,79,24,62,83,44,18,75,41,113,40,151,246,166,179,165,214,223,80,188,97,80,58,155,38,155,47,74,248,20,67,155,5,160,238,98,158,83,194,190,67,77,157,214,254,231,184,219,107,52,77,127,24,102,1,251,144,178,50,149,13,72,65,177,195,218,201,87,134,131,175,32,138,133,5,222,97,231,84,23,145,0,0,0,0,0,0,0,4,41,206,128,242,166,173,199,241,117,198,134,165,146,203,52,95,227,238,179,230,157,217,246,150,222,230,133,245,132,184,119,132,141,103,152,252,95,104,99,219,67,10,130,249,141,182,75,54,161,214,140,241,127,170,234,139,166,120,25,52,114,229,18,143,207,253,104,37,55,211,118,149,102,205,172,243,69,131,177,163,206,196,98,47,43,253,141,83,226,156,196,105,10,77,247,10,236,181,186,106,38,189,26,85,26,251,214,169,108,81,34,175,213,207,232,27,50,12,237,189,170,53,40,169,229,145,66,35,223,91,159,197,71,248,89,5,51,45,26,107,203,217,23,146],[1,0,0,0,0,0,0,0,3,0,0,0,0,0,0,0,145,0,0,0,0,0,0,0,4,152,114,8,226,152,129,189,143,159,235,189,61,126,77,186,177,15,177,198,231,5,34,111,125,133,166,28,66,30,241,227,59,80,219,209,42,163,202,187,2,195,122,112,97,36,57,11,1,211,21,128,2,225,173,117,181,10,180,163,2,95,194,192,197,47,20,11,73,107,64,162,186,132,11,124,214,88,140,12,251,69,62,23,141,83,2,96,21,114,228,19,194,60,10,241,245,51,4,112,163,62,149,80,145,205,86,193,203,69,217,220,173,33,215,247,80,176,64,133,51,142,249,107,142,120,179,55,137,175,200,161,84,102,99,113,62,168,235,217,165,32,55,228,213,145,0,0,0,0,0,0,0,4,221,149,209,31,50,228,196,193,40,117,20,66,130,133,167,86,232,131,255,67,132,106,254,5,143,9,175,168,200,26,164,176,67,148,98,98,222,11,25,3,228,254,142,237,212,73,95,96,110,238,141,127,144,83,49,59,74,232,96,228,191,253,97,216,58,228,9,219,171,28,211,114,147,98,128,186,194,69,37,205,51,171,30,102,54,225,233,116,73,90,187,77,26,248,83,39,96,230,195,183,192,95,211,68,44,117,64,43,181,212,233,145,28,25,84,130,206,91,146,213,29,122,36,213,187,244,203,93,16,77,236,46,253,213,226,70,101,220,4,246,60,200,163,47,145,0,0,0,0,0,0,0,4,118,236,102,255,65,189,12,26,115,11,178,130,105,223,84,158,135,128,9,1,180,170,35,104,224,216,71,114,102,103,251,234,127,217,183,115,2,17,190,231,245,104,142,247,27,216,29,178,181,1,208,166,116,118,186,16,218,36,114,187,196,84,204,246,0,229,191,181,50,130,12,141,47,132,164,18,165,13,106,235,181,171,193,122,62,126,175,113,253,14,26,199,239,142,167,76,99,212,149,132,198,50,10,11,123,211,114,232,137,78,117,130,155,38,27,106,122,219,221,34,141,219,253,91,249,73,134,81,31,134,243,143,240,164,23,114,81,161,77,241,94,52,47,67],[2,0,0,0,0,0,0,0,3,0,0,0,0,0,0,0,145,0,0,0,0,0,0,0,4,241,132,141,26,40,54,65,201,52,48,86,28,100,57,102,106,142,94,36,177,250,127,115,148,189,34,186,71,15,164,119,226,250,197,157,48,40,236,159,63,80,229,161,192,21,53,21,39,32,27,230,41,248,81,122,150,238,78,155,55,225,175,106,169,115,101,103,73,66,119,243,244,229,127,33,76,57,94,222,110,172,161,101,120,37,227,250,45,205,42,233,90,34,105,196,174,132,242,39,90,90,83,153,197,181,36,45,8,89,218,182,138,146,212,39,101,125,63,10,57,52,191,208,142,40,215,184,107,87,172,33,93,128,165,57,129,215,204,52,165,146,81,229,226,145,0,0,0,0,0,0,0,4,94,40,188,81,188,105,191,202,35,228,70,225,229,181,40,243,109,142,213,0,38,93,214,202,139,44,212,42,86,56,184,220,94,141,8,171,175,55,6,9,146,248,79,146,132,170,182,210,12,205,175,86,61,81,99,72,26,53,124,200,8,62,232,145,58,88,140,80,156,182,196,199,13,248,182,135,95,7,234,230,133,254,118,12,132,222,173,124,226,9,36,63,64,63,161,202,183,243,182,118,155,97,51,22,36,34,203,142,155,227,158,30,78,39,10,247,135,207,212,83,207,190,180,242,79,199,46,180,179,183,140,158,220,71,192,35,152,229,130,139,159,203,14,228,145,0,0,0,0,0,0,0,4,229,251,155,218,232,213,74,252,105,57,155,105,170,138,98,149,27,48,37,163,92,41,51,192,123,239,185,132,28,113,210,58,240,253,6,78,210,78,165,20,129,90,61,59,36,164,50,215,134,129,125,125,71,17,122,108,246,237,122,2,236,152,215,214,2,140,47,74,187,242,14,125,150,112,140,180,154,47,91,9,161,142,91,254,245,186,121,120,144,13,38,247,209,191,215,79,204,198,4,58,23,241,86,128,197,147,38,177,243,159,161,104,215,82,153,242,222,127,144,99,171,20,112,43,252,224,102,61,57,224,130,86,32,5,116,56,102,117,104,148,20,96,102,209]]]
-
-
-  //fixing parts.
-  //there is a bug in web3-eth-abi handling byte arrays,
-  //but buffer work fine.
-  const parts = [];
-  partsRaw.forEach(x => {
-    parts.push( Buffer.from(x));
-  });
-
+  const { parts, acks } = testdata.getTestPartNAcks();
 
   // one epoch in 5 seconds day.
   const stakingEpochDuration = new BN(5);
